requirements.txt
Django>=6.0
djangorestframework
joblib
pandas
numpy
scikit-learn
imbalanced-learn
